services/api_gateway/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
httpx==0.27.2
python-jose==3.3.0
pydantic==2.8.2

